+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to the "Semantic AI Commit" extension will be documented in this file.
 
+## v0.1.0 - 05/01/2026
+
+- Enhanced selection for multiple open repositories.
+- Added support for the commit message language.
+
 ## v0.0.3 - 23/09/2025
 
 - Added demo GIFs for SCM menu and Command Palette in the README.